Add the missing contextual search component
