+++ packages/storybook/stories/inputs/search.jsx
@@ -1,17 +1,24 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
-import { InputSearch } from '@equisoft/design-elements-react';
+import { Search, SearchGlobal } from '@equisoft/design-elements-react';
 
 storiesOf('Search Bar', module)
     .add('Contextual', () => (
-        <InputSearch
+        <Search
             id="searchbar_contextual"
             label="Search"
             placeholder="Ex.: Miky Mike"
         />
     ))
+    .add('Global', () => (
+        <SearchGlobal
+            id="searchbar_global"
+            label="Search"
+            placeholder="Ex.: Marquee Mark"
+        />
+    ))
     .add('Disabled', () => (
-        <InputSearch
+        <SearchGlobal
             disabled
             id="searchbar_disabled"
             label="Search"
